requirements.txt
pandas
tqdm
PyGithub
requests
beautifulsoup4
python-dotenv